+++ Cargo.toml
@@ -12,4 +12,4 @@ futures = "0.3.13"
 tokio = { version = "1.4.0", features = [ "sync", "rt-multi-thread", "macros", "process" ] }
 
 [dev-dependencies]
-pseudo = "0.2.0"
+mockall = "0.9.1"
